requirements.txt
# Core dependencies
monai==1.3.0
torch==2.0.0
torchvision==0.15.0

# Data processing
numpy==1.21.0
pandas==1.3.0
scikit-learn==1.0.0
nibabel==3.2.0

# Visualization
matplotlib==3.4.0
seaborn==0.11.0

# Utilities
tqdm==4.62.0
pyyaml==6.0
pytorch-grad-cam==0.6.0
nilearn==0.9.0

# Optional: for Jupyter notebooks
jupyter==1.0.0
ipykernel==6.0.0 